Change collision phsyics to be better, no longer crashes with divide by zero error maybe?? Fruits now bounce off of walls accurately, and bounce off eachother inaccurately.

+++ src/main.rs
@@ -63,7 +63,7 @@ fn main(mut gba: agb::Gba) -> ! {
             //Fruit was just dropped, create new fruit
             //Create position - For now it's a default value, will be set to the player's pos in the future
             let initial_pos: Vector2D<FixedNum<8>> = Vector2D::new((WIDTH/2).into(), (5).into());
-            held_fruit = create_fruit(initial_pos, &oam, fruit_sprites.as_slice(), (fruit_objects.len() as i8)%11, fruit_objects.len() as i32);
+            held_fruit = create_fruit(initial_pos, &oam, fruit_sprites.as_slice(), (fruit_objects.len() as i32)%11, fruit_objects.len() as i32);
             held_fruit.object.show();
         }
